requirements.txt
Django>=5.1
djangorestframework
drf-spectacular
django-filter
django-cors-headers
channels[daphne]
celery
redis
psycopg2-binary
python-dotenv
requests
google-generativeai
openai
azure-storage-blob
Pillow
gunicorn
uvicorn
whitenoise
firebase-admin
